# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Engineering Analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Repositories ─────────────────────────────────────────────────────────
  /repositories:
    get:
      operationId: listRepositories
      tags: [repositories]
      summary: List all repositories
      parameters:
        - name: teamId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repository"
    post:
      operationId: createRepository
      tags: [repositories]
      summary: Register a repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
  /repositories/{id}:
    get:
      operationId: getRepository
      tags: [repositories]
      summary: Get a repository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Repository
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRepository
      tags: [repositories]
      summary: Update a repository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
    delete:
      operationId: deleteRepository
      tags: [repositories]
      summary: Delete a repository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Deployments ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Teams ─────────────────────────────────────────────────────────────
    Team:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        timezone:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TeamInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        timezone:
          type: string
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        timezone:
          type: string

    # ── Repositories ──────────────────────────────────────────────────────
    Repository:
      type: object
      required: [id, name, fullName, createdAt]
      properties:
        id:
          type: integer
        teamId:
          type: ["integer", "null"]
        name:
          type: string
        fullName:
          type: string
        url:
          type: ["string", "null"]
        defaultBranch:
          type: string
        language:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RepositoryInput:
      type: object
      required: [name, fullName]
      properties:
        teamId:
          type: integer
        name:
          type: string
        fullName:
          type: string
        url:
          type: string
        defaultBranch:
          type: string
        language:
          type: string
    RepositoryUpdate:
      type: object
      properties:
        teamId:
          type: integer
        name:
          type: string
        url:
          type: string
        defaultBranch:
          type: string
        language:
          type: string

    # ── Deployments ───────────────────────────────────────────────────────
...