# lib/api-zod/src/generated/types/teamUpdate.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * Engineering Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  timezone?: string;
}